template files for home, login and signup pages
